The default runner config.toml contains:
```
> cat /etc/gitlab-runner/config.toml
concurrent = 1
```

which limits the total number of concurrent jobs among all runners to
one.

This adds a parameter, `concurrent`, to the `gitlab::cirunner` class to
manage this setting.  By default, the setting is left unmanaged, but
the parameter can be set to any integer >= 1 to enable management of it.

Closes #71

Contributed by @iamjamestl. Thanks a lot!