Fix gitlab-ctl reconfigure LD_LIBRARY_PATH warning
This patch ensures LD_LIBRARY_PATH is unset in the `gitlab-ctl configure`
`exec` resource.

As shown in https://travis-ci.org/github/voxpupuli/puppet-gitlab/jobs/739916738
on lines 2089-2090, `gitlab-ctl reconfigure` emits a warning when LD_LIBRARY_PATH
is set. It emits this warning even if that enviroment variable is empty.  Since
LD_LIBRARY_PATH is always set by the Puppet wrapper script used in the `puppet`
CLI, the `exec` that runs `gitlab-ctl configure` must unset this variable to
avoid the warning message.