cirunner: update to use the new runner registration process
Currently we register the runners using a personal token from the admin
account however this method of registration is deprectated and will be
removed in 18.0.  As such we need to move to the newer registration
process.

this MR moves to the newer method and simplifies a lot of the complexity
to support runners we don't use.  long term we should switch to the
voxpupil module.

Bug: DNSIN-183